Simplify module utilities
