+++ mylib/_utils.py
@@ -1,3 +1,4 @@
+import random
 import functools as ft
 from pathlib import Path
 from urllib.parse import ParseResult, urlunparse
@@ -33,3 +34,19 @@ class DatasetPathHandler:
 
     def to_string(self, path):
         return urlunparse(self.to_url(path))
+
+class Backoff:
+    _backoff_factor = 2
+
+    def __init__(self, backoff, fuzz=0):
+        self.backoff = backoff
+        self.fuzz = fuzz
+
+    def __iter__(self):
+        backoff = self.backoff
+        while True:
+            if self.fuzz:
+                backoff += backoff * random.uniform(-self.fuzz, self.fuzz)
+            yield backoff
+
+            backoff *= self._backoff_factor
